AppSchemasController constants moved to top ControllerSelector handles routing for schemas
